Made minor changes to the tests for the RentRepository interface as well as RentService and RentController classes.

+++ src/test/java/com/nikoladronjak/rently/controller/RentControllerTest.java
@@ -90,18 +90,18 @@ class RentControllerTest {
 
 		customer = new Customer(1, "Mika", "Mikic", "mika@gmail.com", "mika123", null);
 
-		officeSpace1 = new OfficeSpace(1, "Office Space 1", "Jove Ilica 154", "", 300, 150, true, 30, photos, owner,
-				null, 100, null);
-		officeSpace2 = new OfficeSpace(2, "Office Space 2", "Studentski trg 1", "", 250, 120, true, 20, photos, owner,
-				null, 90, null);
+		officeSpace1 = new OfficeSpace(1, "Office Space 1", "Jove Ilica 154", "", (double) 300, 150, true, 30, photos,
+				owner, null, 100, null);
+		officeSpace2 = new OfficeSpace(2, "Office Space 2", "Studentski trg 1", "", (double) 250, 120, true, 20, photos,
+				owner, null, 90, null);
 
 		lease1 = new Lease(1, 200, new GregorianCalendar(2024, 11, 12), new GregorianCalendar(2025, 11, 12),
 				officeSpace1, customer, null);
 		lease2 = new Lease(2, 250, new GregorianCalendar(2024, 11, 12), new GregorianCalendar(2025, 11, 12),
 				officeSpace2, customer, null);
 
-		utilityLease1 = new UtilityLease(1, 40, utility1, officeSpace1, rents);
-		utilityLease2 = new UtilityLease(2, 60, utility2, officeSpace1, rents);
+		utilityLease1 = new UtilityLease(1, (double) 40, utility1, officeSpace1, rents);
+		utilityLease2 = new UtilityLease(2, (double) 60, utility2, officeSpace1, rents);
 
 		utilityLeases = new ArrayList<UtilityLease>();
 		utilityLeases.add(utilityLease1);
